+++ fdm-app/app/lib/config.server.ts
@@ -39,6 +39,12 @@ export const serverConfig: ServerConfig = {
         nmi: {
             api_key: String(process.env.NMI_API_KEY),
         },
+        rvo: {
+            clientId: String(process.env.RVO_CLIENT_ID),
+            redirectUri: String(process.env.RVO_REDIRECT_URI),
+            clientName: String(process.env.RVO_CLIENT_NAME),
+            pkioPrivateKey: String(process.env.RVO_PKIO_PRIVATE_KEY),
+        },
         gemini: {
             api_key: String(process.env.GEMINI_API_KEY),
             model: process.env.GEMINI_MODEL,